+++ packages/gatsby-theme/src/components/content.js
@@ -7,29 +7,39 @@ import * as R95Components from '@react95/core';
 import IconRenderer from './icon-renderer';
 import { TASKBAR_HEIGHT } from '../utils/constants';
 
+const ConditionalWrapper = ({ condition, wrapper, children }) =>
+  condition ? wrapper(children) : children;
+
 const Content = ({ content }) => {
   if (!content.body) return null;
 
   const {
     body,
-    frontmatter: { title, icon },
+    frontmatter: { title, icon, modal },
   } = content;
 
   return (
-    <R95Components.Modal
-      title={title}
-      icon={<IconRenderer {...icon} />}
-      style={{
-        top: 0,
-        height: `calc(100% - ${TASKBAR_HEIGHT}px)`,
-        width: '100%',
-      }}
-      closeModal={() => navigate('/')}
+    <ConditionalWrapper
+      condition={modal}
+      wrapper={chidren => (
+        <R95Components.Modal
+          title={title}
+          icon={<IconRenderer {...icon} />}
+          style={{
+            top: 0,
+            height: `calc(100% - ${TASKBAR_HEIGHT}px)`,
+            width: '100%',
+          }}
+          closeModal={() => navigate('/')}
+        >
+          {chidren}
+        </R95Components.Modal>
+      )}
     >
       <MDXProvider components={R95Components}>
-        <MDXRenderer>{body}</MDXRenderer>
+        <MDXRenderer frontmatter={content.frontmatter}>{body}</MDXRenderer>
       </MDXProvider>
-    </R95Components.Modal>
+    </ConditionalWrapper>
   );
 };